duckdb tests
