+++ pyproject.toml
@@ -40,7 +40,8 @@ ignore = ["PLR2004", "PLR0913", "PLR0917"]
 select = ["E", "F", "I", "W", "PL"]
 
 [tool.ruff.lint.per-file-ignores]
-"__init__.py" = ["E402"]
+"__init__.py" = ["E402"]  # module import not at top of file
+"**/tests/*" = ["PLC2701"]  # private ignores
 
 [tool.ruff.lint.isort]
 force-single-line = true
