+++ tests/test_prom_utils.py
@@ -1,7 +1,4 @@
-from datakube.prom_utils import PromReader
-
-
-def test_compute_pod_owners_map() -> None:
-    reader = PromReader("./tests/data")
-    df = reader.compute_pod_owners_map()
-    assert len(df) == 193
+# def test_compute_pod_owners_map() -> None:
+#     reader = PromReader("./tests/data")
+#     df = reader.compute_pod_owners_map()
+#     assert len(df) == 193
